Add files via upload
This method makes use of the fact that with factors a, b of a nonprime number n, the values for a and b cannot both be larger than sqrt(n). So at least one factor has to be less than sqrt(n). If there are no factors between 2 and sqrt(n) then n is prime.